+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Features
 * Add support for arbitrary albumentations transforms `#1001 <https://github.com/azavea/raster-vision/pull/1001>`_
 * Minor tweaks to regression learner `#1013 <https://github.com/azavea/raster-vision/pull/1013>`_
 * Add ability to specify number of PyTorch reader processes `#1008 <https://github.com/azavea/raster-vision/pull/1008>`_
+* Make img_sz specifiable `#1012 <https://github.com/azavea/raster-vision/pull/1012>`_
 
 Bug Fixes
 ~~~~~~~~~~~~